Title: 올바른 등식 만들기, Time: 306ms, Memory: 10MB, Status: Passed - CodeTree

+++ 240619/올바른 등식 만들기/right-equality.java
@@ -10,7 +10,7 @@ public class Main {
             nums[i] = sc.nextInt();
         
         int offset = 20;
-        int[][] dp = new int[101][41];
+        long[][] dp = new long[101][41];
         dp[0][0 + offset] = 1;
         m += offset;
